[addons] create from AddonDll.h a new AddonDll.cpp
Needed to fix linker fault after remove of template usage. Further
are a lot of include passed to the place where it is used.
