+++ xbmc/cores/AudioEngine/Engines/ActiveAE/AudioDSPAddons/ActiveAEDSP.cpp
@@ -49,6 +49,7 @@ extern "C" {
 #include "settings/MediaSourceSettings.h"
 #include "settings/Settings.h"
 #include "utils/JobManager.h"
+#include "utils/log.h"
 #include "utils/StringUtils.h"
