+++ xbmc/peripherals/addons/AddonInputHandling.cpp
@@ -26,6 +26,7 @@
 #include "peripherals/addons/AddonButtonMap.h"
 #include "peripherals/devices/PeripheralJoystick.h"
 #include "peripherals/Peripherals.h"
+#include "utils/log.h"
 
 using namespace JOYSTICK;
 using namespace PERIPHERALS;
