+++ xbmc/pvr/PVRManager.cpp
@@ -63,6 +63,7 @@
 #include "utils/log.h"
 #include "utils/Stopwatch.h"
 #include "utils/StringUtils.h"
+#include "utils/URIUtils.h"
 #include "utils/Variant.h"
 #include "video/VideoDatabase.h"
 #include "ServiceBroker.h"
