+++ xbmc/pvr/timers/PVRTimerInfoTag.cpp
@@ -30,6 +30,7 @@
 #include "settings/Settings.h"
 #include "utils/log.h"
 #include "utils/StringUtils.h"
+#include "utils/URIUtils.h"
 #include "utils/Variant.h"
 
 #include "pvr/PVRManager.h"
