+++ xbmc/pvr/timers/PVRTimers.cpp
@@ -38,6 +38,7 @@
 #include "threads/SingleLock.h"
 #include "utils/log.h"
 #include "utils/StringUtils.h"
+#include "utils/URIUtils.h"
 #include "utils/Variant.h"
 
 using namespace PVR;
