+++ xbmc/pvr/windows/GUIWindowPVRGuide.cpp
@@ -31,7 +31,7 @@
 #include "settings/Settings.h"
 #include "threads/SingleLock.h"
 #include "utils/log.h"
-
+#include "utils/StringUtils.h"
 #include "pvr/PVRGUIActions.h"
 #include "pvr/PVRManager.h"
 #include "pvr/addons/PVRClients.h"
